+++ packages/moxxmpp/lib/src/xeps/xep_0444.dart
@@ -8,8 +8,8 @@ import 'package:moxxmpp/src/stanza.dart';
 import 'package:moxxmpp/src/stringxml.dart';
 import 'package:moxxmpp/src/util/typed_map.dart';
 
-class MessageReactions implements StanzaHandlerExtension {
-  const MessageReactions(this.messageId, this.emojis);
+class MessageReactionsData implements StanzaHandlerExtension {
+  const MessageReactionsData(this.messageId, this.emojis);
   final String messageId;
   final List<String> emojis;
 
@@ -59,7 +59,7 @@ class MessageReactionsManager extends XmppManagerBase {
         message.firstTag('reactions', xmlns: messageReactionsXmlns)!;
     return state
       ..extensions.set(
-        MessageReactions(
+        MessageReactionsData(
           reactionsElement.attributes['id']! as String,
           reactionsElement.children
               .where((c) => c.tag == 'reaction')
@@ -72,7 +72,7 @@ class MessageReactionsManager extends XmppManagerBase {
   List<XMLNode> _messageSendingCallback(
     TypedMap<StanzaHandlerExtension> extensions,
   ) {
-    final data = extensions.get<MessageReactions>();
+    final data = extensions.get<MessageReactionsData>();
     return data != null
         ? [
             data.toXML(),