+++ packages/moxxmpp/test/xeps/xep_0334_test.dart
@@ -1,5 +1,4 @@
 import 'package:moxxmpp/moxxmpp.dart';
-import 'package:moxxmpp/src/util/typed_map.dart';
 import 'package:test/test.dart';
 
 import '../helpers/manager.dart';