+++ packages/moxxmpp/CHANGELOG.md
@@ -15,6 +15,7 @@
 - `MessageManager` now allows registering callbacks for adding data whenever a message is sent.
 - **BREAKING**: `MessageEvent` now makes use of `TypedMap`.
 - **BREAKING**: Removed `PresenceReceivedEvent`. Use a manager registering handlers with priority greater than `[PresenceManager.presenceHandlerPriority]` instead.
+- **BREAKING**: `ChatState.toString()` is now `ChatState.toName()`
 
 ## 0.3.1